Rename the module to DBIx::RoboQuery and limit its focus to that
There was no need to reserve a generic namespace (DBIx::Enabler)
for things yet to come.
There aren't any ideas for further modules to be put in the namespace,
and frankly if it isn't related to this Query object then
why should it be in the same distribution/namespace?

So limit the focus of this module/distribution/repo
to what it was originally designed for (query/results)
and rename the modules to reflect that.